Create s2geometryapi.cc
Command line version.  server.cc with all the server stuff stripped out so that it can be called from the command line.  The use case is an Apache server with php.  Php scripts call the s2geometryapi executable with the URL arguments on the command line.  The executable writes the response body to stdout and returns the HTML response code as an error code (so rather than the usual exit(0) meaning no error, exit(200) means no error).  It's a hack.  It's inefficient.  It's a very quick way to get a server up and running as the front end of a C++ library.